+++ ui/apps/platform/src/Containers/Vulnerabilities/VulnerablityReporting/utils.test.ts
@@ -1,4 +1,4 @@
-import { ReportStatus } from 'services/ReportsService.types';
+import { ReportStatus } from 'types/reportJob';
 import { getReportStatusText, getCVEsDiscoveredSinceText } from './utils';
 import { ReportParametersFormValues } from './forms/useReportFormValues';
